[Update] extracted the exercise card into a reuseable component
